requirements.txt
pymongo==4.11
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.1
python-telegram==0.19.0
python-telegram-bot==20.3
telebot==0.0.5
telegram==0.0.1
